# [AGENT] Improve http v1 parse performance