--- a/agent/benches/flow_generator/flow_map.rs
+++ b/agent/benches/flow_generator/flow_map.rs
@@ -14,24 +14,18 @@
  * limitations under the License.
  */
 
-use std::sync::Arc;
-use std::time::Duration;
-use std::time::Instant;
+use std::time::{Duration, Instant};
 
 use criterion::*;
 
 use deepflow_agent::{
-    _FlowPerfCounter as FlowPerfCounter, _TcpFlags as TcpFlags, _TcpPerf as TcpPerf,
-    _benchmark_report as benchmark_report,
-    _benchmark_session_peer_seq_no_assert as benchmark_session_peer_seq_no_assert,
-    _meta_flow_perf_update as meta_flow_perf_update,
-    _new_flow_map_and_receiver as new_flow_map_and_receiver, _new_meta_packet as new_meta_packet,
-    _reverse_meta_packet as reverse_meta_packet,
+    _TcpFlags as TcpFlags, _new_flow_map_and_receiver as new_flow_map_and_receiver,
+    _new_meta_packet as new_meta_packet, _reverse_meta_packet as reverse_meta_packet,
 };
 
 use public::proto::common::TridentType;
 
-fn bench_flow_map(c: &mut Criterion) {
+pub(super) fn bench(c: &mut Criterion) {
     c.bench_function("flow_map_syn_flood", |b| {
         b.iter_custom(|iters| {
             let (mut map, _) = new_flow_map_and_receiver(TridentType::TtProcess);
@@ -94,39 +88,3 @@ fn bench_flow_map(c: &mut Criterion) {
         })
     });
 }
-
-fn bench_perf(c: &mut Criterion) {
-    c.bench_function("perf_stats_report", |b| {
-        b.iter_custom(|iters| {
-            let mut perf = TcpPerf::new(Arc::new(FlowPerfCounter::default()));
-            let start = Instant::now();
-            for _ in 0..iters {
-                benchmark_report(&mut perf);
-            }
-            start.elapsed()
-        })
-    });
-    c.bench_function("perf_update", |b| {
-        b.iter_custom(|iters| {
-            let mut perf = TcpPerf::new(Arc::new(FlowPerfCounter::default()));
-            let start = Instant::now();
-            for _ in 0..iters {
-                meta_flow_perf_update(&mut perf);
-            }
-            start.elapsed()
-        })
-    });
-    c.bench_function("perf_session_peer_seq_no_assert_desc", |b| {
-        b.iter(|| {
-            benchmark_session_peer_seq_no_assert(true);
-        })
-    });
-    c.bench_function("perf_session_peer_seq_no_assert", |b| {
-        b.iter(|| {
-            benchmark_session_peer_seq_no_assert(false);
-        })
-    });
-}
-
-criterion_group!(benches, bench_flow_map, bench_perf);
-criterion_main!(benches);