--- a/agent/src/flow_generator/perf/mod.rs
+++ b/agent/src/flow_generator/perf/mod.rs
@@ -15,7 +15,7 @@
  */
 
 mod dns;
-mod http;
+pub mod http;
 pub mod l7_rrt;
 mod mq;
 mod rpc;